+++ api/ilos-validator/src/index.ts
@@ -1,0 +1,3 @@
+export { AjvValidatorProvider } from './AjvValidatorProvider';
+export { ValidatorMiddleware } from './ValidatorMiddleware';
+export { ValidatorProviderInterface, ValidatorProviderInterfaceResolver } from './ValidatorProviderInterface';
