+++ api/providers/validator/src/index.ts
@@ -1,3 +1,4 @@
+export { ValidatorProviderInterface, ValidatorProviderInterfaceResolver } from '@ilos/provider-validator';
+
 export { ValidatorProvider } from './ValidatorProvider';
 export { ValidatorMiddleware } from './ValidatorMiddleware';
-export { ValidatorProviderInterface, ValidatorProviderInterfaceResolver } from './ValidatorProviderInterface';
