+++ api/proxy/package.json
@@ -14,7 +14,7 @@
     "coverage": "export NODE_ENV=testing; nyc --all --reporter=text mocha src/**/*.spec.ts tests/**/*.spec.ts",
     "test": "export NODE_ENV=testing; mocha src/**/*.spec.ts",
     "test:integration": "export NODE_ENV=testing; mocha tests/**/*.spec.ts",
-    "lint": "tslint -c tslint.json -p tsconfig.json",
+    "lint": "tslint -c ../tslint.json -p tsconfig.json",
     "debug": "yarn build && yarn watch-debug",
     "serve-debug": "nodemon --inspect dist/index.js",
     "watch-debug": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"cyan.bold,green.bold\" \"yarn watch-ts\" \"yarn watch-node\""
