+++ api/proxy/src/middlewares/signResponseMiddleware.ts
@@ -3,9 +3,6 @@ import expressMung from 'express-mung';
 
 /**
  * Hash the payload
- *
- * @param payload
- * @returns {*}
  */
 const hashPayload = (payload) => {
   const hash = crypto.createHash('SHA256');
@@ -15,13 +12,7 @@ const hashPayload = (payload) => {
 
 /**
  * Nest response body in a payload and hash it
- *
- * @param body
- * @param req
- * @param res
- * @returns {{public_key: (*|string), signature: *, payload: *}}
  */
-// eslint-disable-next-line no-unused-vars
 const signResponse = (body, req, res) => {
   const signed = {
     sha256: hashPayload(body),
